fix POST /books

+++ src/main/java/pl/akademiaqa/bos/books/api/payload/UpdateBookPayload.java
@@ -21,5 +21,5 @@ public class UpdateBookPayload {
     BigDecimal price;
     @NotNull
     @PositiveOrZero
-    Long available;
+    Integer available;
 }
